# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards Management API
  description: API for managing bank cards with admin functionality
  contact:
    name: API Support
    email: api@example.com
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /cards/block/{id}:
    patch:
      tags:
        - Cards Controller
      summary: Request card blocking
      description: Request blocking of a specific card. The card status will be changed
        to BLOCK_REQUESTED for admin review
      operationId: requestCardBlocking
      parameters:
        - name: id
          in: path
          description: ID of the card to block
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Blocking request submitted successfully
          content:
            application/json:
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_LONG
                  - 414 REQUEST_URI_TOO_LONG
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
        "404":
          description: Card not found
          content:
            application/json:
              example:
                error: Card not found
                timestamp: "1757615223938"
        "403":
          description: Access denied - card belongs to another user
          content:
            application/json:
              example:
                error: You do not have permission to access this card
                timestamp: "1757615223938"
...
  /cards/{id}:
    get:
      tags:
        - Cards Controller
      summary: Get card details
      description: Retrieve detailed information about a specific card owned by the
        authenticated user
      operationId: getCard
      parameters:
        - name: id
          in: path
          description: ID of the card to retrieve
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Card retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
              example:
                id: 1
                ownerId: 2
                cardNumber: '**** **** **** 5678'
                expiryDate: 2029-09-11 15:05:03.693
                status: ACTIVE
                balance: 174.88
        "404":
          description: Card not found
          content:
            application/json:
              example:
                error: Card not found
                timestamp: "1757615223938"
        "403":
          description: Access denied - card belongs to another user
          content:
            application/json:
              example:
                error: You do not have permission to access this card
                timestamp: "1757615223938"
  /cards/{id}/card-number:
    get:
      tags:
        - Cards Controller
      summary: Get full card number
      description: Retrieve the complete (decrypted) card number for a specific card.
        Requires authentication.
      operationId: getCardNumber
      parameters:
        - name: id
          in: path
          description: ID of the card to get number for
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Card number retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardNumberResponse"
              example:
                cardNumber: "1234567812345678"
        "404":
          description: Card not found
          content:
            application/json:
              example:
                error: Card not found
                timestamp: "1757615223938"
        "403":
          description: Access denied - card belongs to another user
          content:
            application/json:
              example:
                error: You do not have permission to access this card
                timestamp: "1757615223938"
  /cards/{id}/balance:
    get:
      tags:
        - Cards Controller
      summary: Get card balance
      description: Retrieve the current balance of a specific card
      operationId: getBalance
      parameters:
        - name: id
          in: path
          description: ID of the card to get balance for
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Balance retrieved successfully
          content:
            application/json:
              example: 1500.75
        "404":
          description: Card not found
          content:
            application/json:
              example:
                error: Card not found
                timestamp: "1757615223938"
        "403":
          description: Access denied - card belongs to another user
          content:
            application/json:
              example:
                error: You do not have permission to access this card
                timestamp: "1757615223938"
  /cards/cards:
    get:
      tags:
        - Cards Controller
      summary: Get user cards with filtering
      description: Retrieve paginated list of user's cards with sorting and filtering
        options
      operationId: getCards
      parameters:
        - name: page
          in: query
          description: Page number (zero-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          example: 0
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
          example: 10
        - name: sort
          in: query
          description: "Sorting criteria in format: property:direction. Multiple sorts\
          \ supported"
          required: false
          schema:
            type: array
            default:
              - balance:desc
            items:
              type: string
        - name: status
          in: query
          description: Filter by card status
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
              - BLOCK_REQUESTED
          example: ACTIVE
        - name: minBalance
          in: query
          description: Minimum balance filter
          required: false
          schema:
            type: number
          example: 100.0
        - name: maxBalance
          in: query
          description: Maximum balance filter
          required: false
          schema:
            type: number
          example: 5000.0
      responses:
        "200":
          description: Cards retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponse"
              example:
                content:
                  - id: 1
                    ownerId: 2
                    cardNumber: '**** **** **** 5678'
                    expiryDate: 2029-09-11 15:05:03.693
                    status: ACTIVE
                    balance: 673.24
                page: 0
                limit: 10
                totalElements: 3
                totalPages: 1
  /admin/users/{userId}/cards:
    get:
      tags:
        - Admin Users Controller
      summary: Get user cards
      description: Retrieves paginated list of user cards with sorting and filtering
        options
      operationId: getUserCards
      parameters:
        - name: userId
          in: path
          description: ID of the user to retrieve cards for
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: page
          in: query
          description: Page number (zero-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          example: 0
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
          example: 10
        - name: sort
          in: query
          description: "Sorting criteria in format: property:direction. Multiple sorts\
          \ supported"
          required: false
          schema:
            type: array
            default:
              - balance:desc
            items:
              type: string
        - name: status
          in: query
          description: Filter by card status
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - BLOCKED
              - EXPIRED
              - BLOCK_REQUESTED
          example: ACTIVE
        - name: minBalance
          in: query
          description: Minimum balance filter
          required: false
          schema:
            type: number
          example: 100.0
        - name: maxBalance
          in: query
          description: Maximum balance filter
          required: false
          schema:
            type: number
          example: 5000.0
      responses:
        "200":
          description: Cards retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponse"
              example:
                content:
                  - id: 1
                    ownerId: 2
                    cardNumber: '**** **** **** 5678'
                    expiryDate: 2029-09-11 15:05:03.693
                    status: ACTIVE
                    balance: 673.24
                page: 0
                limit: 10
                totalElements: 3
                totalPages: 1
        "403":
          description: Access denied - admin privileges required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseCardResponse"
...
components:
  schemas:
...
    CardResponse:
      type: object
      description: Card response DTO containing card details
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the card
          example: 1
        ownerId:
          type: integer
          format: int64
          description: Identifier of the card owner
          example: 2
        cardNumber:
          type: string
          description: Masked card number (16 digits)
          example: '**** **** **** 5678'
        expiryDate:
          type: string
          description: Expiry date in YYYY-MM-DD HH:MM:SS.SSS format
          example: 2029-09-11 15:05:03.693
        status:
          type: string
          description: Current status of the card
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
            - BLOCK_REQUESTED
          example: ACTIVE
        balance:
          type: number
          description: Current balance of the card
          example: 174.15
    CardNumberResponse:
      type: object
      description: Card number response DTO containing decrypted card number
      properties:
        cardNumber:
          type: string
          description: Decrypted card number (16 digits)
          example: 1234567812345678
    PageResponse:
      type: object
      description: Paginated response wrapper
      properties:
        content:
          type: array
          description: List of items on the current page
          items: {}
        page:
          type: integer
          format: int32
          description: Current page number (zero-based)
          example: 0
        limit:
          type: integer
          format: int32
          description: Number of items per page
          example: 10
        totalElements:
          type: integer
          format: int64
          description: Total number of elements across all pages
          example: 100
        totalPages:
          type: integer
          format: int32
          description: Total number of pages
          example: 10
    PageResponseCardResponse:
      type: object
      description: Paginated response wrapper
      properties:
        content:
          type: array
          description: List of items on the current page
          items:
            $ref: "#/components/schemas/CardResponse"
        page:
          type: integer
          format: int32
          description: Current page number (zero-based)
          example: 0
        limit:
          type: integer
          format: int32
          description: Number of items per page
          example: 10
        totalElements:
          type: integer
          format: int64
          description: Total number of elements across all pages
          example: 100
        totalPages:
          type: integer
          format: int32
          description: Total number of pages
          example: 10
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT